refactor: utils + hooks 통합
