+++ src/advanced/components/SelectItemSection/CartItem.tsx
@@ -1,6 +1,5 @@
 import React, { useCallback } from 'react';
 import { CartItemType } from '../../types';
-import { generateProductName, generatePriceHtml } from '../../utils/priceUtils';
 
 export interface CartItemPropsType {
   item: CartItemType;
