+++ src/advanced/hooks/index.ts
@@ -1,4 +1,5 @@
-export * from './useCart';
-export * from './useProducts';
-export * from './useTimers';
-export * from './useManual';
+export { useCart } from './useCart';
+export { useProducts } from './useProducts';
+export { useTimers } from './useTimers';
+export { useManual } from './useManual';
+export { useDiscount } from './useDiscount';
